+++ Changes.md
@@ -4,6 +4,7 @@
 - Fixes #20: All temporary headers values are strings.
 - Reduced cache storage consumption by only storing response body and headers.
 - Increased maximum cache content length size to 1.25 MB.
+- You may pass a redis url to the :cache option of setup.
 
 ### 1.6.1
 - NEW: Batch requests are now parallelized.